pyproject.toml
[tool.poetry]
name = "auto_analyst"
version = "0.1.0"
description = ""
authors = ["shtmjp <shtm.1058@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"
discord-webhook = "^1.3.1"
oauth2client = "^4.1.3"
pandas = "^2.2.3"
tqdm = "^4.67.0"
google = "^3.0.0"
scikit-learn = "^1.5.2"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
bs4 = "^0.0.2"
yfinance = "^0.2.51"
langchain = "^0.3.13"
langchain-google-vertexai = "^2.0.9"
langchain-google-genai = "^2.0.7"
mplfinance = "^0.12.10b0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.7.2"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
